refactor `fetchAllTransactions`
