+++ main/wallet.js
@@ -11,8 +11,6 @@ const { WalletBackend } = require('./wallet-backend')
 /** @typedef {import('./typings').Context} Context */
 /** @typedef {import('./typings').FILTransaction} FILTransaction */
 /** @typedef {import('./typings').FILTransactionProcessing} FILTransactionProcessing */
-/** @typedef {import('./typings').FILTransactionLoading} FILTransactionLoading */
-/** @typedef {import('./typings').TransactionStatus} TransactionStatus */
 
 const log = electronLog.scope('wallet')
 const walletStore = new Store({
@@ -21,12 +19,17 @@ const walletStore = new Store({
 
 const backend = new WalletBackend()
 backend.transactions = loadStoredEntries()
+backend.onTransactionUpdate = function () {
+  walletStore.set('transactions', backend.transactions)
+  sendTransactionsToUI()
+}
+backend.onTransactionSucceeded = async function () {
+  await updateBalance()
+}
 
 /** @type {Context | null} */
 let ctx = null
 let balance = loadBalance()
-/** @type {Set<string>} */
-const stateReplaysBeingFetched = new Set()
 
 /**
  * @param {Context} _ctx
@@ -58,7 +61,7 @@ async function refreshState () {
     log.error('Updating balance', err)
   }
   try {
-    await updateTransactions()
+    await backend.fetchAllTransactions()
   } catch (err) {
     log.error('Updating transactions', err)
   }
@@ -76,105 +79,6 @@ async function updateBalance () {
   balance = await backend.fetchBalance()
   walletStore.set('balance', balance.toFil())
   ctx.balanceUpdate(balance.toFil())
-}
-
-/**
- * @returns {Promise<void>}
- */
-async function updateTransactions () {
-  assert(backend.address)
-
-  console.log(new Date(), 'updateTransactions')
-
-  // Load messages
-  const messages = await backend.getMessages(backend.address)
-
-  // Convert messages to transactions (loading)
-  /** @type {FILTransactionLoading[]} */
-  const transactionsLoading = messages.map(message => {
-    return {
-      height: message.height,
-      hash: message.cid,
-      outgoing: message.from.robust === backend.address,
-      amount: new FilecoinNumber(message.value, 'attofil').toFil(),
-      address: message.from.robust === backend.address
-        ? message.to.robust
-        : message.from.robust
-    }
-  })
-
-  /** @type {(FILTransaction|FILTransactionProcessing)[]} */
-  const updatedTransactions = []
-
-  for (const transactionProcessing of transactionsLoading) {
-    // Find matching transaction
-    const tx = backend.transactions.find(tx => tx.hash === transactionProcessing.hash)
-
-    // Complete already loaded data
-    // Keep references alive by prefering `tx`
-    const transaction = tx || transactionProcessing
-
-    if (!transaction.timestamp && transaction.height) {
-      transaction.timestamp =
-        (await backend.getTipset(transaction.height)).minTimestamp * 1000
-    }
-
-    if (
-      transaction.hash &&
-      (!transaction.status || transaction.status === 'processing') &&
-      !stateReplaysBeingFetched.has(transaction.hash)
-    ) {
-      const { hash } = transaction
-      stateReplaysBeingFetched.add(hash)
-      transaction.status = 'processing'
-      ;(async () => {
-        while (true) {
-          try {
-            const stateReplay = await backend.getStateReplay(hash)
-            transaction.status = stateReplay.receipt.exitCode === 0
-              ? 'succeeded'
-              : 'failed'
-            stateReplaysBeingFetched.delete(hash)
-            walletStore.set('transactions', updatedTransactions)
-            if (transaction.status === 'succeeded') {
-              try {
-                await updateBalance()
-              } catch {}
-            }
-            sendTransactionsToUI()
-            break
-          } catch (err) {
-            console.error(
-              `Failed getting status for ${transactionProcessing.hash}`
-            )
-            await timers.setTimeout(12_000)
-          }
-        }
-      })()
-    }
-
-    updatedTransactions.push(
-      /** @type {FILTransaction} */
-      (transaction)
-    )
-  }
-
-  // Add transaction potentially not yet returned by the API
-  for (const transaction of backend.transactions) {
-    if (!updatedTransactions.find(tx => tx.hash === transaction.hash)) {
-      updatedTransactions.push(transaction)
-      break
-    }
-  }
-
-  // Update state
-  backend.transactions = updatedTransactions
-
-  // Save transactions
-  walletStore.set('transactions', updatedTransactions)
-
-  // Send transaction state to UI
-  sendTransactionsToUI()
 }
 
 function listTransactions () {
